+++ android/src/main/java/br/com/rsmarques/flutter_branch_sdk/FlutterBranchSdkPlugin.java
@@ -465,6 +465,14 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
                 event.addContentItems(buo).logEvent(context);
             }
         });
+        /*
+        new Thread() {
+            @Override
+            public void run() {
+                event.addContentItems(buo).logEvent(context);
+            }
+        }.start();
+         */
     }
 
     private void trackContentWithoutBuo(MethodCall call) {
@@ -481,6 +489,14 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
                 event.logEvent(context);
             }
         });
+        /*
+        new Thread() {
+            @Override
+            public void run() {
+                event.logEvent(context);
+            }
+        }.start();
+         */
     }
 
     private void setIdentity(MethodCall call) {
@@ -496,7 +512,6 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
                 Branch.getAutoInstance(context).setIdentity(userId);
             }
         });
-
     }
 
     private void setRequestMetadata(MethodCall call) {